fix show comments error
