+++ src/scripts/index.js
@@ -3,7 +3,7 @@ import 'bootstrap/dist/css/bootstrap.min.css';
 import '../scss/style.scss';
 import { showsList } from './UX/cards';
 import showPopup from './UX/popup';
-import commentAPI from './API/Involvement';
+import commentAPI from './API/Comments';
 import newCounter from './UX/counter';
 
 showsList.renderCards();
